requirements.txt
anyio==3.3.4
APScheduler==3.6.3
asgiref==3.4.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
fastapi==0.70.0
h11==0.12.0
idna==3.3
numpy==1.21.4
pandas==1.3.4
pydantic==1.8.2
python-dateutil==2.8.2
python-telegram-bot==13.8
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
tornado==6.1
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
uvicorn==0.15.0
